Allow for the message to be parsed as HTML
This allows passing in a message along the lines of `Something <br /> <u>Foo Bar</u>.` and have it render correctly.
Added prop `html`, defaults to safe rendering.
Passing `false` to (or omitting) the `html` prop will function as it currently does.
Passing `true` to the `html` prop will render `message` as HTML.